Added Cpu struct
Moving away from opcode generation and moving towards decoding the
opcodes at runtime. There aren't many actual instructions, and
generating the opcodes was taking a lot of work to generate a lot of
duplicate code. Implementation is going much faster this way, and if it
proves to be a bottleneck, it should be possible to write more
specialized functions later on that dont have to do as much runtime
decoding.